requirements.txt
# CLI
rich==12.0.*
typer==0.4.*
rich-click==1.2.*

# when time calc
arrow==1.2.*
python-dateutil==2.8.*
tzdata
backports.zoneinfo
pydantic==1.9.*
airportsdata
tzlocal==4.*
